Add files via upload
For this project, the Lent network has been implemented using a custom-written convolution layer, and then at the end of each convolution layer, the MaxPoling filters of each layer have been displayed using the Metplot Lib library. And based on this, the changes of the input image are displayed after the operations performed on the image in each layer.
Of course, I checked this with the convolution layer that I wrote before, now I will implement it with this new convolution layer and check it carefully and plot the output feature maps of each layer.